requirements.txt
pandas>=2.2.0
plotly>=6.0.0
matplotlib>=3.10.0
numpy>=2.2.0
Pillow>=11.1.0
scikit-learn>=1.6.0
SQLAlchemy>=2.0.38
psycopg2-binary>=2.9.10
fastapi>=0.116.0
uvicorn>=0.35.0
python-multipart>=0.0.20
PyJWT>=2.10.1
bcrypt>=4.0.0
google-auth>=2.38.0
requests>=2.32.0
httpx>=0.28.0
transformers>=4.48.0
open_clip_torch>=2.30.0
faiss-cpu>=1.10.0
